chore: move bulk_upsert to helpers
